chore(deps): bump toolchain to 1.97.1, un-pin AWS/schemars/shlex, serde_yaml -> serde_norway (CCT-783)
- rust-toolchain.toml 1.90.0 -> 1.97.1 + the four rust Dockerfiles in lockstep
- drop the four AWS =pins (needed rust >= 1.94.1), float to current
- schemars to workspace 1.2, shlex un-pinned, serde_yaml replaced by serde_norway on the crdgen path
- absorb 1.97 clippy fallout; account-name validation rewrite covered by a new unit test

+++ crates/cctui-proto/src/chunk.rs
@@ -228,7 +228,7 @@ mod tests {
         ] {
             let payload = pattern(len);
             let frames = split(&payload).expect("must split");
-            assert_eq!(frames.len() as usize, len.div_ceil(CHUNK_SIZE));
+            assert_eq!(frames.len(), len.div_ceil(CHUNK_SIZE));
             assert_eq!(reassemble(&frames), payload, "roundtrip failed at len={len}");
         }
     }
